requirements.txt
annotated-types==0.6.0
anyio==4.3.0
black==24.3.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
dynaconf==3.2.5
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
mypy-extensions==1.0.0
openai==1.16.2
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
sniffio==1.3.1
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.11.0